multi-node/wazuh-proxy/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
slowapi==0.1.9
PyYAML==6.0.1
watchfiles==0.21.0
